Ensure builtin binaries are aligned to ibf_header
The `builtin_binary` buffers are cast to `ibf_header` later, but may not
be aligned to the necessary boundaries (in this case, 4 bytes).
